requirements.txt
hexdump
pycryptodome